Fix linking error
